requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aistudio-sdk==0.3.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
async-timeout==4.0.3
attrs==25.4.0
audioread==3.1.0
av==16.0.1
azure-core==1.36.0
azure-identity==1.25.1
bce-python-sdk==0.9.60
brotli==1.2.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorlog==6.10.1
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
dashscope==1.25.1
dataclasses-json==0.6.7
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
easyocr==1.7.2
exceptiongroup==1.3.0
fastapi==0.128.0
ffmpy==1.0.0
filelock==3.20.0
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2025.10.0
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.48.0
google-auth-httplib2==0.2.1
google-genai==1.63.0
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
gradio==6.3.0
gradio_client==2.0.3
groovy==0.1.2
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.1.5
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
imagesize==1.4.1
isodate==0.7.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==1.1.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.0
langchain-openai==1.1.0
langchain-text-splitters==1.0.0
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.10
langsmith==0.4.47
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.45.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mdurl==0.1.2
modelscope==1.34.0
moviepy==2.2.1
mpmath==1.3.0
msal==1.34.0
msal-extensions==1.3.1
msgpack==1.1.2
multidict==6.7.0
mypy_extensions==1.1.0
nano-vectordb==0.0.4.3
networkx==3.4.2
ninja==1.13.0
numba==0.62.1
numpy==2.2.6
openai==2.8.1
opencv-contrib-python==4.10.0.84
opencv-python==4.12.0.88
opencv-python-headless==4.11.0.86
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
paddleocr==3.4.0
paddlex==3.4.1
pandas==2.3.3
pillow==11.3.0
platformdirs==4.5.0
pooch==1.8.2
prettytable==3.17.0
proglog==0.1.12
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.2.2
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.4.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
pypdfium2==5.4.0
PySocks==1.7.1
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.2
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rsa==4.9.1
ruamel.yaml==0.19.1
safehttpx==0.1.7
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
semantic-version==2.10.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
socks==0
soundfile==0.13.1
soxr==1.0.0
SQLAlchemy==2.0.44
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.12.0
tomlkit==0.13.3
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
typer==0.21.1
typer-slim==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.40.0
wcwidth==0.6.0
websocket-client==1.9.0
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
